refactor(runner): split scheduler and scheduler tests
