+++ packages/runner/test/scheduler-core.test.ts
@@ -1,25 +1,27 @@
-// Core scheduler tests: basic scheduling, event handling, reactive retries,
-// and stream event success callbacks.
+// Core scheduler tests: basic push-mode scheduling behavior.
 
-import { afterEach, beforeEach, describe, it } from "@std/testing/bdd";
-import { expect } from "@std/expect";
-import { assertSpyCall, assertSpyCalls, spy } from "@std/testing/mock";
-import { storedCfcMetadataAppliesToPath } from "../src/cfc/metadata.ts";
-import { type IExtendedStorageTransaction } from "../src/storage/interface.ts";
-import { Runtime } from "../src/runtime.ts";
 import {
-  type Action,
-  type EventHandler,
+  afterEach,
+  assertSpyCall,
+  assertSpyCalls,
+  beforeEach,
+  describe,
+  expect,
   ignoreReadForScheduling,
+  it,
+  Runtime,
+  signer,
+  space,
+  spy,
+  StorageManager,
+  storedCfcMetadataAppliesToPath,
+  toMemorySpaceAddress,
   txToReactivityLog,
-} from "../src/scheduler.ts";
-import { Identity } from "@commonfabric/identity";
-import { StorageManager } from "@commonfabric/runner/storage/cache.deno";
-import type { Entity } from "@commonfabric/memory/interface";
-import { toMemorySpaceAddress } from "../src/link-utils.ts";
-
-const signer = await Identity.fromPassphrase("test operator");
-const space = signer.did();
+} from "./scheduler-core-test-utils.ts";
+import type {
+  Action,
+  IExtendedStorageTransaction,
+} from "./scheduler-core-test-utils.ts";
 
 describe("scheduler", () => {
   let storageManager: ReturnType<typeof StorageManager.emulate>;
@@ -1386,517 +1388,4 @@ describe("scheduler", () => {
     await resultCell.pull();
     expect(runCount).toBe(2);
   });
-});
-
-describe("event handling", () => {
-  let storageManager: ReturnType<typeof StorageManager.emulate>;
-  let runtime: Runtime;
-  let tx: IExtendedStorageTransaction;
-
-  beforeEach(() => {
-    storageManager = StorageManager.emulate({ as: signer });
-    // Create runtime with the shared storage provider
-    // We need to bypass the URL-based configuration for this test
-    runtime = new Runtime({
-      apiUrl: new URL(import.meta.url),
-      storageManager,
-    });
-    // Use push mode for event handling tests
-    runtime.scheduler.disablePullMode();
-    tx = runtime.edit();
-  });
-
-  afterEach(async () => {
-    await tx.commit();
-    await runtime?.dispose();
-    await storageManager?.close();
-  });
-
-  it("should queue and process events", async () => {
-    const eventCell = runtime.getCell<number>(
-      space,
-      "should queue and process events 1",
-      undefined,
-      tx,
-    );
-    eventCell.set(0);
-    const eventResultCell = runtime.getCell<number>(
-      space,
-      "should queue and process events 2",
-      undefined,
-      tx,
-    );
-    eventResultCell.set(0);
-    tx.commit();
-
-    let eventCount = 0;
-
-    const eventHandler: EventHandler = (tx, event) => {
-      eventCount++;
-      eventResultCell.withTx(tx).send(event);
-    };
-
-    runtime.scheduler.addEventHandler(
-      eventHandler,
-      eventCell.getAsNormalizedFullLink(),
-    );
-
-    runtime.scheduler.queueEvent(eventCell.getAsNormalizedFullLink(), 1);
-    runtime.scheduler.queueEvent(eventCell.getAsNormalizedFullLink(), 2);
-
-    await eventResultCell.pull();
-
-    expect(eventCount).toBe(2);
-    expect(eventCell.get()).toBe(0); // Events are _not_ written to cell
-    expect(eventResultCell.get()).toBe(2);
-  });
-
-  it("should remove event handlers", async () => {
-    const eventCell = runtime.getCell<number>(
-      space,
-      "should remove event handlers 1",
-      undefined,
-      tx,
-    );
-    eventCell.set(0);
-    tx.commit();
-
-    let eventCount = 0;
-
-    const eventHandler: EventHandler = (tx, event) => {
-      eventCount++;
-      eventCell.withTx(tx).send(event);
-    };
-
-    const removeHandler = runtime.scheduler.addEventHandler(
-      eventHandler,
-      eventCell.getAsNormalizedFullLink(),
-    );
-
-    runtime.scheduler.queueEvent(eventCell.getAsNormalizedFullLink(), 1);
-    await eventCell.pull();
-
-    expect(eventCount).toBe(1);
-    expect(eventCell.get()).toBe(1);
-
-    removeHandler();
-
-    runtime.scheduler.queueEvent(eventCell.getAsNormalizedFullLink(), 2);
-    await eventCell.pull();
-
-    expect(eventCount).toBe(1);
-    expect(eventCell.get()).toBe(1);
-  });
-
-  it("should handle events with nested paths", async () => {
-    const parentCell = runtime.getCell<{ child: { value: number } }>(
-      space,
-      "should handle events with nested paths 1",
-      undefined,
-      tx,
-    );
-    parentCell.set({ child: { value: 0 } });
-    tx.commit();
-
-    let eventCount = 0;
-
-    const eventHandler: EventHandler = () => {
-      eventCount++;
-    };
-
-    runtime.scheduler.addEventHandler(
-      eventHandler,
-      parentCell.key("child").key("value").getAsNormalizedFullLink(),
-    );
-
-    runtime.scheduler.queueEvent(
-      parentCell.key("child").key("value").getAsNormalizedFullLink(),
-      42,
-    );
-    await runtime.idle();
-
-    expect(eventCount).toBe(1);
-  });
-
-  it("should process events in order", async () => {
-    const eventCell = runtime.getCell<number>(
-      space,
-      "should process events in order 1",
-      undefined,
-      tx,
-    );
-    eventCell.set(0);
-    tx.commit();
-
-    const events: number[] = [];
-
-    const eventHandler: EventHandler = (_tx, event) => {
-      events.push(event);
-    };
-
-    runtime.scheduler.addEventHandler(
-      eventHandler,
-      eventCell.getAsNormalizedFullLink(),
-    );
-
-    runtime.scheduler.queueEvent(eventCell.getAsNormalizedFullLink(), 1);
-    runtime.scheduler.queueEvent(eventCell.getAsNormalizedFullLink(), 2);
-    runtime.scheduler.queueEvent(eventCell.getAsNormalizedFullLink(), 3);
-
-    await runtime.idle();
-
-    expect(events).toEqual([1, 2, 3]);
-  });
-
-  it("does not commit or flush outbox effects when event handler throws", async () => {
-    const eventCell = runtime.getCell<number>(
-      space,
-      "does not commit or flush outbox effects when event handler throws 1",
-      undefined,
-      tx,
-    );
-    eventCell.set(0);
-    const resultCell = runtime.getCell<number>(
-      space,
-      "does not commit or flush outbox effects when event handler throws 2",
-      undefined,
-      tx,
-    );
-    resultCell.set(0);
-    await tx.commit();
-
-    let attempts = 0;
-    let errors = 0;
-    let flushed = 0;
-    runtime.scheduler.onError(() => {
-      errors++;
-    });
-
-    const eventHandler: EventHandler = (handlerTx) => {
-      attempts++;
-      resultCell.withTx(handlerTx).send(1);
-      handlerTx.enqueuePostCommitEffect({
-        id: "event-handler-throw-outbox",
-        kind: "test",
-        flush() {
-          flushed++;
-        },
-      });
-      throw new Error("boom");
-    };
-
-    runtime.scheduler.addEventHandler(
-      eventHandler,
-      eventCell.getAsNormalizedFullLink(),
-    );
-
-    runtime.scheduler.queueEvent(eventCell.getAsNormalizedFullLink(), 1);
-    await runtime.idle();
-    await runtime.idle();
-
-    expect(attempts).toBe(1);
-    expect(errors).toBe(1);
-    expect(flushed).toBe(0);
-    expect(resultCell.get()).toBe(0);
-  });
-
-  it("should trigger recomputation of dependent cells", async () => {
-    const eventCell = runtime.getCell<number>(
-      space,
-      "should trigger recomputation of dependent cells 1",
-      undefined,
-      tx,
-    );
-    eventCell.set(0);
-    const eventResultCell = runtime.getCell<number>(
-      space,
-      "should trigger recomputation of dependent cells 2",
-      undefined,
-      tx,
-    );
-    eventResultCell.set(0);
-    tx.commit();
-
-    let eventCount = 0;
-    let actionCount = 0;
-    let lastEventSeen = 0;
-
-    const eventHandler: EventHandler = (tx, event) => {
-      eventCount++;
-      eventResultCell.withTx(tx).send(event);
-    };
-
-    const action = (tx: IExtendedStorageTransaction) => {
-      actionCount++;
-      lastEventSeen = eventResultCell.withTx(tx).get();
-    };
-    runtime.scheduler.subscribe(action, {
-      reads: [],
-      shallowReads: [],
-      writes: [],
-    }, {});
-    await eventResultCell.pull();
-
-    runtime.scheduler.addEventHandler(
-      eventHandler,
-      eventCell.getAsNormalizedFullLink(),
-    );
-
-    expect(actionCount).toBe(1);
-
-    runtime.scheduler.queueEvent(eventCell.getAsNormalizedFullLink(), 1);
-    await eventResultCell.pull();
-
-    expect(eventCount).toBe(1);
-    expect(eventResultCell.get()).toBe(1);
-
-    expect(actionCount).toBe(2);
-
-    runtime.scheduler.queueEvent(eventCell.getAsNormalizedFullLink(), 2);
-    await eventResultCell.pull();
-
-    expect(eventCount).toBe(2);
-    expect(eventResultCell.get()).toBe(2);
-    expect(actionCount).toBe(3);
-    expect(lastEventSeen).toBe(2);
-  });
-
-  it(
-    "should retry event handler when the handler transaction aborts, up to retries count",
-    async () => {
-      const entityId = `test:retry-conflict-${Date.now()}` as Entity;
-
-      // Set up an event cell and commit initial state
-      const eventCell = runtime.getCell<number>(
-        space,
-        "should retry event handler on conflict",
-        undefined,
-        tx,
-      );
-      eventCell.set(0);
-      await tx.commit();
-
-      // Event handler that writes a conflicting value to the same entity
-      let attempts = 0;
-      const handler: EventHandler = (tx, _event) => {
-        attempts++;
-        // Force commit failure for the first 5 attempts to exercise retries.
-        if (attempts <= 5) {
-          tx.abort("force-abort-for-retry");
-          return;
-        }
-        // On the final attempt, perform a regular write.
-        tx.write({
-          space,
-          id: entityId,
-          path: [],
-        }, { version: 2 });
-      };
-
-      runtime.scheduler.addEventHandler(
-        handler,
-        eventCell.getAsNormalizedFullLink(),
-      );
-
-      // Queue event (uses default retries configured in scheduler)
-      runtime.scheduler.queueEvent(
-        eventCell.getAsNormalizedFullLink(),
-        1,
-      );
-
-      // First idle may return before the commit callback schedules retries.
-      await runtime.idle();
-      // Wait for any re-queued events to process.
-      await runtime.idle();
-
-      // Should attempt initial + default retries times (DEFAULT_RETRIES=5)
-      expect(attempts).toBe(6);
-
-      // No further assertions needed; this verifies retry behavior only.
-    },
-  );
-});
-
-describe("reactive retries", () => {
-  let storageManager: ReturnType<typeof StorageManager.emulate>;
-  let runtime: Runtime;
-  let tx: IExtendedStorageTransaction;
-
-  beforeEach(() => {
-    storageManager = StorageManager.emulate({ as: signer });
-    runtime = new Runtime({
-      apiUrl: new URL(import.meta.url),
-      storageManager,
-    });
-    // Use push mode for reactive retry tests
-    runtime.scheduler.disablePullMode();
-    tx = runtime.edit();
-  });
-
-  afterEach(async () => {
-    await tx.commit();
-    await runtime?.dispose();
-    await storageManager?.close();
-  });
-
-  it(
-    "should retry reactive actions when commit fails, up to limit",
-    async () => {
-      // Establish a source cell to create a read dependency
-      const source = runtime.getCell<number>(
-        space,
-        "should retry reactive actions when commit fails, up to limit 1",
-        undefined,
-        tx,
-      );
-      source.set(1);
-      await tx.commit();
-      tx = runtime.edit();
-
-      // Count runs; force commit failure each time
-      let attempts = 0;
-      const reactiveAction: Action = (actionTx) => {
-        attempts++;
-        // Read to establish dependency so later changes re-trigger
-        source.withTx(actionTx).get();
-        // Force commit to fail so scheduler retries
-        actionTx.abort("force-abort-for-reactive-retry");
-      };
-
-      // Subscribe and run immediately
-      runtime.scheduler.subscribe(
-        reactiveAction,
-        { reads: [], shallowReads: [], writes: [] },
-        { isEffect: true },
-      );
-
-      // Allow retries to process. Idle may resolve before re-queue occurs,
-      // so loop a few times until attempts reach the expected amount.
-      for (let i = 0; i < 20 && attempts < 10; i++) {
-        await runtime.idle();
-      }
-
-      // MAX_RETRIES_FOR_REACTIVE is 10; expect initial + retries == 10 attempts
-      expect(attempts).toBe(10);
-
-      // After reaching retry limit, a subsequent input change should re-trigger
-      source.withTx(tx).send(2);
-      await tx.commit();
-      tx = runtime.edit();
-
-      // Wait for the follow-up run
-      await runtime.idle();
-
-      expect(attempts).toBe(11);
-    },
-  );
-
-  it(
-    "should preserve dependencies when retrying failed commits",
-    async () => {
-      // This test documents expected behavior for the conflict storm fix:
-      // When a reactive action's commit fails and it retries, it should
-      // preserve its dependency information (not overwrite with empty deps).
-      // This ensures topological sorting works correctly during retries.
-      //
-      // NOTE: This test passes with both buggy and fixed code because line 274
-      // immediately re-learns dependencies after each action run, masking the
-      // bug in simple scenarios. The real bug manifests only in high-concurrency
-      // scenarios (30+ reactive cells) where async commit callbacks race with
-      // scheduler execution. See budget-planner integration test for evidence
-      // of the fix (conflict storm: 65k errors → 1 error after fix).
-
-      const source = runtime.getCell<number>(
-        space,
-        "should preserve dependencies source",
-        undefined,
-        tx,
-      );
-      source.set(1);
-
-      const intermediate = runtime.getCell<number>(
-        space,
-        "should preserve dependencies intermediate",
-        undefined,
-        tx,
-      );
-      intermediate.set(0);
-
-      const output = runtime.getCell<number>(
-        space,
-        "should preserve dependencies output",
-        undefined,
-        tx,
-      );
-      output.set(0);
-
-      await tx.commit();
-      tx = runtime.edit();
-
-      let action1Attempts = 0;
-      let action2Attempts = 0;
-      const action2Values: number[] = [];
-
-      // Action 1: reads source, writes intermediate (will fail first 2 times)
-      const action1: Action = (actionTx) => {
-        action1Attempts++;
-        const val = source.withTx(actionTx).get();
-        intermediate.withTx(actionTx).send(val * 10);
-
-        // Force abort for first 2 attempts to trigger retry logic
-        if (action1Attempts <= 2) {
-          actionTx.abort("force-abort-action1");
-        }
-      };
-
-      // Action 2: reads intermediate, writes output (depends on action1)
-      const action2: Action = (actionTx) => {
-        action2Attempts++;
-        const val = intermediate.withTx(actionTx).get();
-        action2Values.push(val);
-        output.withTx(actionTx).send(val + 5);
-      };
-
-      // Subscribe both actions with correct dependencies
-      runtime.scheduler.subscribe(
-        action1,
-        {
-          reads: [toMemorySpaceAddress(source.getAsNormalizedFullLink())],
-          shallowReads: [],
-          writes: [
-            toMemorySpaceAddress(intermediate.getAsNormalizedFullLink()),
-          ],
-        },
-        {},
-      );
-      runtime.scheduler.subscribe(
-        action2,
-        {
-          reads: [toMemorySpaceAddress(intermediate.getAsNormalizedFullLink())],
-          shallowReads: [],
-          writes: [toMemorySpaceAddress(output.getAsNormalizedFullLink())],
-        },
-        {},
-      );
-
-      // Allow all actions to complete (action1 will retry twice)
-      for (let i = 0; i < 20 && action1Attempts < 3; i++) {
-        await output.pull();
-      }
-
-      // Verify action1 ran 3 times (2 aborts + 1 success)
-      expect(action1Attempts).toBe(3);
-
-      // Action2 should run twice in reactive system:
-      // 1. Initially when both actions run (sees intermediate=0 since action1 aborts)
-      // 2. After action1 succeeds and updates intermediate (sees intermediate=10)
-      expect(action2Attempts).toBe(2);
-      expect(action2Values).toEqual([0, 10]);
-
-      // Critical assertion: The final state must be correct, proving that
-      // dependencies were preserved during retries and topological sort worked.
-      expect(intermediate.get()).toBe(10); // 1 * 10
-      expect(output.get()).toBe(15); // 10 + 5
-    },
-  );
 });
